move IdentifierImpl to internal
