+++ sql/core/src/main/scala/org/apache/spark/sql/catalyst/analysis/ResolveSessionCatalog.scala
@@ -210,6 +210,9 @@ class ResolveSessionCatalog(val catalogManager: CatalogManager)
         c
       }
 
+    case DropTable(ResolvedV1Identifier(ident), ifExists, purge) if conf.useV1Command =>
+      DropTableCommand(ident, ifExists, isView = false, purge = purge)
+
     // v1 DROP TABLE supports temp view.
     case DropTable(ResolvedIdentifier(FakeSystemCatalog, ident), _, _) =>
       DropTempViewCommand(ident)
